tree cutting

+++ cosmetic.go
@@ -282,16 +282,16 @@ func (s *Shoes) InteractOptions() []string {
 	return nil
 }
 
-func (h *Hat) Interact(x int, y int, player *Player, zone *Zone, opt int) {
+func (h *Hat) Interact(x uint8, y uint8, player *Player, zone *Zone, opt int) {
 }
 
-func (s *Shirt) Interact(x int, y int, player *Player, zone *Zone, opt int) {
+func (s *Shirt) Interact(x uint8, y uint8, player *Player, zone *Zone, opt int) {
 }
 
-func (p *Pants) Interact(x int, y int, player *Player, zone *Zone, opt int) {
+func (p *Pants) Interact(x uint8, y uint8, player *Player, zone *Zone, opt int) {
 }
 
-func (s *Shoes) Interact(x int, y int, player *Player, zone *Zone, opt int) {
+func (s *Shoes) Interact(x uint8, y uint8, player *Player, zone *Zone, opt int) {
 }
 
 func (h *Hat) IsItem() {}
